Static linking (#8)
* NEW: Static linking

Initial version, triplet resolution is not automatic.

* FIX: Proper triplet selection

Since scikit framework does not support cmake presets yet,
we need to create something different.
The main flag to define static/dynamic linking is BUILD_SHARED_LIBS.
Based on this we adjust the vcpkg triplet in the CMake.

* TWEAK: Enable windows sanity

With the support of static linking we should be able to run on windows as well.
The issue before was related to legacy.dll dependency,
which in this case will be embedded inside the .pyd module.

* FIX: Packaging

Default triplet override only for windows.
Unify suffix naming.
Package dynamic libs only if linking dynamically.

* TWEAK: Update vanillapdf

Update to main, latest release, supporting static linking